# packages/product-service/openapi.yml
openapi: 3.0.0
components:
  schemas:
    NotFoundResponse:
      properties:
        message:
          type: string
          default: Not found
        status:
          type: int32
          default: 404
...
    ProductResponse:
      properties:
        productId:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
          default: 0
          minimum: 0
  securitySchemes: {}
info:
  title: product-service
  description: Product service API
  version: '1'
paths:
...
  '/product/{id}':
    get:
      operationId: getProductById
      summary: Get product by product ID
      description: Provide productId as a path params
      responses:
        '200':
          description: Found product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'